9种消息有16位的消息标识

+++ NewLife.MQTT/Messaging/MqttIdMessage.cs
@@ -8,7 +8,7 @@ namespace NewLife.MQTT.Messaging
     {
         #region 属性
         /// <summary>标识</summary>
-        public Int32 Id { get; set; }
+        public UInt16 Id { get; set; }
         #endregion
 
         #region 读写方法
